requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
tavily-python
wikipedia
trustcall
langgraph-cli[inmem]
utils
feedparser
html2text
markdownify
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
xqlatm==0.1.3     # (supuesto nombre del package que quieres usar)
python-dotenv==1.0.1
psycopg2-binary